Code/prompt_sentiment/llama_cpp/requirements.txt
python-dotenv==1.0.0
huggingface_hub
llama-cpp-python[server]